// types/pdp.d.ts
export type ProductAttributes = {
  UPC: string;
  MODEL_CODE: string;
  COLOR_CODE: string;
  TITLE: string;
  AGE_GROUP: "Adult" | "Kids" | "Teens" | string;
  GENDER: "Male" | "Female" | "Unisex" | string;
  MARKETING_THEME: string;
  COLLECTION_MARKETING_THEME_STORY: string;
  PERSONA: string;
  LAUNCH_DATE: string; // ISO date string (YYYY-MM-DD)
  RIM_TYPE: string;
  FRAME_SHAPE: string;
  FRAME_MATERIAL: string;
  FRONT_FRAME_COLOR: string;
  FRONT_MATERIAL: string;
  FRONT_MATERIAL_FAMILY: string;
  TEMPLE_MATERIAL: string;
  TEMPLE_MATERIAL_FAMILY: string;
  FLEX_HINGE: "Yes" | "No" | string;
  A_BOX: number;
  DBL: number;
  TEMPLE_LENGTH: number;
  HINGE_DISTANCE: number;
  SIZE: "XS" | "S" | "M" | "L" | "XL" | string;
  FRONT_COLOR_FAMILY: string;
  TEMPLE_COLOR: string;
  LENS_TYPE: "Eyeglasses" | "Sunglasses" | string;
  LENS_COLOR_DESCRIPTION: string;
  POLARIZED: string;
  SUSTAINABLE_FRONT: string;
  SUSTAINABLE_TEMPLE: string;
  STATUS_N1_26: string;
  SRP: number;
  PICTURE: string;
  DISTRO_N1_26_PINS: number;
  PERCENT_DISTRO_N1_26: number;
  DISTRO_CLASSIFICATION_AT_MODEL_LEVEL: string;
  WEBSITE_SELECTION_AVRIL_2026: "Yes" | "No" | string;
  NPI_BS_ADV: string;
  NPI_BS: string;
  COLLECTION: string;
  CAMPAIGN_FOOTER_SELECTION: string;
};